Move grid helpers to /helpers
